LayoutTests/fast/scrolling/latching/latched-scroll-in-passive-region-expected.txt

CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
CONSOLE MESSAGE: Unable to preventDefault() on event of type 'wheel': Inside passive event listener.
Tests that a gesture in the passive wheel event region dispatches more than one wheel event

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS wheelEventHandlerCount > 4 is true
PASS successfullyParsed is true

TEST COMPLETE
